pyproject.toml
[project]
name = "langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=1.2.0",
    "langchain-groq>=1.1.1",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "streamlit>=1.52.2",
]
